+++ src/common.h
@@ -58,7 +58,7 @@ JxlEncoderFrameSettings* frameConfigToJxlEncoderFrameSettings(
     const FrameConfig& frameConfig,
     uint32_t tpsNumerator, uint32_t tpsDenominator,
     uint32_t frameXsize, uint32_t frameYsize,
-    const std::optional<int32_t>& brotliEffort = {});
+    const std::optional<int16_t>& brotliEffort = {});
 
 /**
  * Return (index,info) pairs for all non-JXL-reserved ISO BMFF boxes.
